Add Thread Pooling as a Threading Policy (#100)
This allows changing the threading model from a per request model to a pooled model.  

- Threading Policies
-- Pooled
This uses a queue, and thread pool to execute requests to Kinesis.  This limits the number of threads that the native process may use.  Under extremely heavy load this can increase latency significantly more than the per request model.  

-- Per Request
This is the current default for 0.12.x versions of the KPL, and creates a thread for each request to Kinesis.  Under heavy load this can create a very large number of threads, which may cause resource exhaustion.

Updated compiler for Linux from GCC 4.9 to GCC 6.3.