Add A Larger Default Set of Certificates (#154)
* Update Certificates, and Add an Extractor for Them

Add a full set of root CA certificates, and added a certificate
extractor to extract them to the correct location.

* Add a Global Fixture for the AWS SDK

Added a global fixture that will initialize the AWS SDK before any
tests run.  This should ensure that all tests that interact with the
SDK succeed.

* Use the Certificate Extractor Directory as the CA Directory

Use the directory returned by the certificate extractor as the CA
directory passed to the native application.

* Remove Explicit Set of C++ standard

Removed the explicit set of the C++ standard option, and instead use
the CMake variable.